+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add new matcher `toHaveReceivedAnyCommand`
+
 ## [4.0.1] - 2024-10-16
 
 ### Changed